Use JS mock API when hybrid testing
The parserTests.js script uses service-runner as a library to launch the
mockAPI.js as a worker and all workers listen on the same port.  It then
goes on to run the tests and eventually gets to spawnSync in hybrid
testing, which is a blocking operations.

From the php side, a curl request is made to the mockAPI, at which
point, in node.js default configuration, the master will accept the
connection and hand off the handle to the worker.  But, the master is
blocked above on the sync spawn.

Node.js has another mode where workers can accept connections directly.
This can be enabled with the env var `NODE_CLUSTER_SCHED_POLICY=none`.

Bug: T225235
Change-Id: Ia67da7db7141f10b61ba55d615057a20f38c9708