Add REST API endpoints for wt->html
Adds the following endpoints:
* /{domain}/v3/page/{format}/{title}
* /{domain}/v3/page/{format}/{title}/{revision}
* /{domain}/v3/transform/{from}/to/{format}
* /{domain}/v3/transform/{from}/to/{format}/{title}
* /{domain}/v3/transform/{from}/to/{format}/{title}/{revision}

Only wt->html parameter combinations are accepted, and validation
and error handling is done extremely poorly (it will be redone
when I3c37ea2b lands). The layout of the code is messy but
relatively easy to match with the JS equivalent. That will also
be redone once the REST API has its own validation framework.

To test, you need something like I0f9aaa9c to be able to access
the API from the web.

Bug: T221169
Bug: T221170
Change-Id: I65d10e7a27e4d87544125152e7ff2ee6a3492cd9