WikiLinkHandler.buildLinkAttrs always returns an KV[] for contentKVs
Assert the type of the result and remove some dead code which thought
that somehow the value could be a string instead of a KV.  Also fix the
dereferences in the PHP port since this is a KV not an associative array.
Finally, avoid creating a DOM Fragment token if we don't have to (this
also helps ensure that the srcOffsets of a DOM Fragment are always non-null
in a future patch).

Also renamed $captionTk, which confusingly suggested that the value was
a token.  It's just an associative array; used the name $optsCaption to
make it clear this is just an alias for $opts['caption'];

Change-Id: I352054e15efa93dad0e3a2e848340bcd69433fb1