Validate @covers tags and fix all the broken ones
PHPUnit requires that @covers tags are well-formed, otherwise it will
immediately fail and not generate any coverage. They must be fully
qualified (full namespace) and don't respect namespace/use statements.

The best way to do this going forwards is to have a @coversDefaultClass
at the top of the test, and then use @covers ::funcName for each
individual test case. I've done that for all the tests that were broken.

And covers-validator will now run as part of CI to prevent future
regressions.

Bug: T226276
Change-Id: I246c5e011ee5f55e6a9bad32efadf539d0b71999