Fix OOM when parsing template
By extending source similar to the js side.

With a bonus sanitizer fix.

What seems to be happening is that although the `$extSrc` shouldn't
exceed the input length as written, because `strlen( $extSrc )`
can be less than the start position `$dp->tsr->start`, the combination
of the two can exceed it so that the `$skipPos` is off the end.

Presumably, this puts the tokenizer in a sad state.

Bug: T226451
Change-Id: I858fadb29686b5ec5294f8f712d95eb37ceb70ac