Convert `extTagOffsets` to a DomSourceRange
The extTagOffsets was its own funny thing, sort of like a KVSourceRange
but using 'key' and 'value' for the start and end tag.  Convert it to
use the standard DomSourceRange format.

Even though this changes data-parsoid, the extTagOffsets property in
data-parsoid isn't actually used by any current html2wt code, so this
is safe to deploy without a version bump.

Change-Id: Ib8eabedaba50937f915d28a0a3d234eb6f5efdeb