Pass `frame` instead of `env` in DOM processors and handlers
RT testing found a failure in:

 $ echo -e "<gallery>File:Foo.jpg|caption</center></gallery>" | bin/parse.js --wt2wt --rtTestMode
 <gallery>
 File:Foo.jpg|caption></galler
 </gallery>

Where the text after `caption` should be `</center>` but
ProcessTreeBuilderFixups was using the wrong source text when it
tried to add the `mw:Placeholder/StrippedTag` marker.  This is a
follow up to 31d356a5147939bab889e190faa254563f160e06 (ensure
proper source texts) although it was exposed most directly by
2853a90ceda7cdfae1512ac1969df93a72561768 which changed the
source text that the Gallery extension used.

Add a `setFrame` implementation to DOMPostProcessor so that it
gets the correct frame, and then pass this to the various DOM processors
as part of the options object/array.  Since you can get the `env` from
the `frame`, in most places it makes sense just to replace the `env`
parameter to various methods with `frame`.  Always use `frame.srcText`
instead of `env.topFrame.srcText`, which is likely to be wrong.

Change-Id: Ied6f6ac3a945a6bae62d005d024342695bf528d0