Fix regression in [[:ru:Fable Legends]]
This page on ruwiki would cause a crash:

 $ bin/parse.js --domain ru.wikipedia.org --pageName Fable_Legends < /dev/null
 [...eventually...]
{"name":"parsoid","hostname":"earth","pid":11021,"level":60,"logType":"fatal","wiki":"ruwiki","title":"Fable Legends","oldId":null,"reqId":null,"userAgent":"Roundtrip-Test","msg":"Cannot read property 'ownerDocument' of undefined","stack":"TypeError: Cannot read property 'ownerDocument' of undefined
    at insertElement (/home/subbu/work/wmf/deploy/node_modules/domino/lib/HTMLParser.js:2645:30)
    at insertHTMLElement (/home/subbu/work/wmf/deploy/node_modules/domino/lib/HTMLParser.js:2620:15)
    at in_body_mode (/home/subbu/work/wmf/deploy/node_modules/domino/lib/HTMLParser.js:5723:9)
    at Object.insertToken (/home/subbu/work/wmf/deploy/node_modules/domino/lib/HTMLParser.js:2521:7)
    at HTML5TreeBuilder.processToken (/home/subbu/work/wmf/parsoid/lib/wt2html/HTML5TreeBuilder.js:226:17)
    at HTML5TreeBuilder.onChunk (/home/subbu/work/wmf/parsoid/lib/wt2html/HTML5TreeBuilder.js:103:9)
    at SyncTokenTransformManager.emitter.addListener.tokens (/home/subbu/work/wmf/parsoid/lib/wt2html/HTML5TreeBuilder.js:57:47)
 [...]

The source of this crasher was an extra EOFTk which was being emitted, which
was causing the document to be closed; inserting the next token after the
EOFTk was causing this crash.  Root cause was a bug in
d5a16abb7dd1c513a12e62010562d7eb5602e3f7 where I didn't convert the sense
of the test properly; using proper names for the DSR offsets (as we do in
the PHP port) should make these bugs less likely in the future!

Change-Id: I47c4744f6438b084524c455df4bfef7f8b99c36f