Optimize regexp used to verify UTF-8 validity
PHP7 was hitting the PREG_JIT_STACKLIMIT_ERROR on large strings here,
because the `*` quantifier was implicitly growing the stack for each
character to allow backtracking if the match failed.  Backtracking
is pointless here, so use `*+` to eliminate it (and its stack usage).

While we're add it, add the `D` modifier to ensure that `$` only matches
at the very end of the string.  Given that a trailing newline would
presumably be valid UTF-8, this doesn't actually affect correctness.

Change-Id: I6244b18de41b829c1983dbc19894e169ededa443