Remove parserTests-whitelist
This was used before we had separate html/php and html/parsoid sections.

Use the modern construct, purge and remove all trace of the old thing.

Transfer some useful comments from the whitelist to parserTests.txt.

The T4702 test is added to the html2html blacklist because we were
previously using a bogus HTML clause which didn't match the wikitext,
but which happened to roundtrip correctly.  Fixing the wt2html test
caused an html2html failure due to escaping added in html2wt.

Change-Id: I4a5b488e29041f9f818d84a9cf8bab6414e480f1