Assert KVs aren't found when calling TokenUtils::tokensToString()
We are calling this with KV->v, which has a slightly different signature
but don't expect the mismatched portion to crop up.

This also clarifies that the signature for KV->v can also include an
array of a mixture of Tokens and strings.  A purely array<string> should
probably have already been joined before getting here though.

Bug: T228223
Change-Id: I6dee4d89721e3869c27c06bb4a01ac7ac462ca6f