Fix dsr test to check start against null, not 'truthy'
This code is triggered on wikitext with badly nested <a> tags, such
as [http://example.org [[Foo]]]

The test *should* be that the start dsr is non-null; '0' is a legit
value.

However, this fix is mostly theoretical, as I can't figure out how to
write a test case such that dsr start is actually 0 here: you need
opening brackets in wikitext to start the link, and those will always
bump the dsr up by at least one.  Still, I'm pretty confident this
patch is the *correct* way to do this test.

Follow up to a comment on I7795cedf14e6ff56a31eeaba0a32c3c5c3166f08.

Change-Id: I059a17412f03c8dfc15589d90bf85d1848021b00