Backport changetree modifications from 7f8d2fc
In 7f8d2fc, changetrees were ensured to always be arrays, which results
in some slightly different trees.  Here we backport those changes to the
JS side so that we have matching test generation / blacklists.

A check is added to the recursive return value of genChangesInternal so
that empty arrays are reverted to the default change type.

This also fixes comparison type mismatch in TestUtils.php

Change-Id: I49a3024c706d3f3b233632b67c8688f6bbe2e09e