Restrict the use of the dummy env to title parsing
Using it to create docs means that those docs will remain live while the
entire test suite runs, since the env holds a reference to them,
bloating memory and slowing things down.

It also avoids using a document with an environment which didn't create
it, preventing any future pitfalls with that.

At some point we should think about decoupling title parsing from the
environment, but that's for another day.

Unfortunately, there's some ugly manipulation of the page content that's
necessary for invoking the various modes with the same environment
instance.  In the future, these parsoid internals should be opaque to
the test runner and it should only calls methods exposed by the
entrypoint.

Change-Id: Ife4709c6aa15c6301dd5eab2e91e5f2dd832ee67