Ignore blacklisted dsr info
Since these aren't equivalent between JS and PHP, this is just showing a
bunch of false negative.  Ignore it for now so we can concentrate on
what's actually broken.  But, remember to remove it when we switch to a
PHP output blacklist.

This fixes about 40 tests with the default modes.

Change-Id: I88d665309bfd8dc1dd7fc2f934488c174932cf71