Don't defeat the MW parser's internal caching
Instead of calling Parser::preprocess() every time a template is seen,
call replaceVariables() instead. Call Parser::clearState() only when a
new document is started, as is traditional.

This provides a 39% reduction in parse time on the [[Aam Aadmi Party]]
test case in local testing.

Compared to the JS version, the problem is the lack of an API response
cache, which is not compensated for anywhere in the stack, so repeated
template invocations lead to parsing those templates again from scratch.

Change-Id: If15723a71e61987137afe0df8fc9aff086ad372b