Backport link escaping change from 9822f75
In the test "<pre> inside a link", `target` is an array that stringifies
as "Main,[object Object]" from,

[ 'Main',
  SelfclosingTagTk {
    name: 'extension',
    attribs: [ [Object], [Object], [Object], [Object], [Object] ],
    dataAttribs:
     { tsr: [Object],
       stx: 'html',
       src: '<pre> Page</pre>',
       extTagOffsets: [Object] } } ]

This passed the valid link target check and failed the valid protocol
check, so would indicate that escaping *was* required, differing from
returning `false` early on the php side.

Update the blacklist so that these can match.

Change-Id: Id57073de0637e1c5dd7e46f88f61f7125eb474fa