Implement REST API endpoint for wikitext to lint transform
Implemented REST API endpoint for linting wikitext, which was
previously stubbed.

Modified Parsoid::wikitext2html() return type declaration to
allow returning either a PageBundle or an array of linting
results. This is not an ideal long-term solution, but given
that the entry point interface otherwise does not have a way
to return linting results, it is a functional temporary solution

Added a linting test case to ParsoidTest.

Bug: T221175
Change-Id: I7d20b9dc016f0290bf9f6a82d150cfef1861f4cc