Avoid a O(N^2) blowup when assigning pagebundle ids
The id attribute is not indexed in the PHP DOM by default.  This caused
an O(N^2) blowup.  This workaround could still fail if nodes in the
pre-pagebundle graph have IDs starting with `mw`, but this should get
most cases working.  A proper fix probably requires moving to a better
PHP DOM implementation (T217867).

Bug: T232180
Change-Id: Ie37f489783bf1b8d906ac30913eaa3d46056b4b0