Fix selser in the REST API
Previously, if wikitext was supplied, it was ignored.  If it wasn't,
then selser was disabled.

Same caveat as in the cli, the oldText in the Selser class is just for
show. The serializer gets its wikitext from $env->pageConfig not from 
the Selser class. This will be cleaned up in a future patch.

Change-Id: Id03ff950547e11350b79d03409433dfa7c18c855