Ensure template frame has correct source text
This was a porting bug; the JavaScript equivalent file already passes
the `srcText` option, but it was overlooked in the PHP port.  As a result,
the child frame created for the template expansion has the parent's
source text, not the template source text.  This DSR substring operations
are garbled and eventually UTF-8 invalidity checks are triggered.

Added test case to verify that inner frame has the correct title.

Change-Id: I908ab756cdc583ec01a057827e7a1b52a4f34db5