Handle case-sensitivity in behavior switches correctly in Parsoid/JS
JavaScript doesn't have embedded case-sensitivity flags in regular
expressions (eg, `(?i: ... )` ), and so we kinda hacked around the
issue in JavaScript, recognizing the all-uppercase and all-lowercase
versions of behavior switches and then sometimes (but not always!)
throwing a global 'i' flag on the entire regular expression at the
end.

PHP has embedded case-sensitivity flags, so we can do this "correctly"
in the PHP port.  However, first quickly hack together "better"
support on the JavaScript side so we can check for any behavior or
parser test changes (and not incorrectly blame them on porting bugs).

Change-Id: If32c948fca00933d1ad21fa580615cb438029aec