Use PHPUtils::lastItem() over end() in more places
Tim Starling has indicated in couple different places that end(..)
is not preferred and he had implemented a private version of lastItem
in the PEG grammar code whereas PHPUtils::lastItem was recommending
use of end(..).

In this patch, I moved the implementation from the grammar to PHPUtils
and replaced end(..) with PHPUtils::lastItem in a number of places in
the codebase. We should discuss whether we want to use this helper
everywhere.

Resolved a couple of PORT-FIXMEs in the bargain.

Change-Id: I837f2a98003df8ab7dbdf9af045e17bdd6e27799