Optionally convert offsets in generated HTML
This adds a new --offsetType option to parse.php as well as a new query
option to the Parsoid API to allow post-conversion of offsets to/from either
UCS-2 or unicode codepoints ("char") from the native UTF-8 ("byte") offsets
we use internally.

Initially this is used in the round-trip testing framework to emit offsets
which are compatible with the output of Parsoid/JS.  Longer term, either
we'll officially expose this, or deprecate and remove it once we're done
testing against Parsoid/JS output.

To avoid confusion, we record the offsetType used in the top-level pagebundle
object, and check this during html2wt conversions to make sure we're not
mixing up DSR types.

Bug: T222738
Bug: T232056
Change-Id: Id04eb332190b84b6b4b412daaffeca9ea509b001