Fix regression in round trip testing script.
Apparently the redirectToOldId REST API method didn't cooperate well
with the mechansism we were using to ensure parse.php --html2wt didn't
crash because of the absence of a revision.

Fixed, but there's a more fundamental issue here with ensure srcText
is always set correctly in the Env.  Filed T234549 as a follow up.

Bug: T233818
Change-Id: I41320af7c726ea33d57442f90cd44fd970b2b6d9