Use SourceRange objects in Html2Wt serializer code
This brings us one step closer to unifying SerializerState::getOrigSrc()
and WTUtils::getWTSource().

This also fixes a minor issue in Separators.php, where we constructed an
ad-hoc range by adding or subtracting 1 *byte* to an offset (instead of
one *character*) which could create a malformed UTF-8 string.  The
subsequent preg_match would then fail on that string because it was bad
UTF-8.  This still yielded the correct results, since we were only
returning the substring if it matched space or tab, but by a surprising
path.

Bug: T234548
Change-Id: Ia4332b61f5809835017725c84dd921f02ee0a708