Add --offsetType option to parserTests.php
This ensures we exercise the bidirectional DSR offset conversion
in the wt2html and html2wt modes of parserTests, and should catch
any obvious crashers.  (But probably wouldn't detect if we simply
failed to do any conversion in either direction.)

The selser mode ought to detect offset conversion bugs.
Note that we disable blacklist literal string comparison, though,
since the blacklist is always stored with embedded UTF-8 offsets.
(We still use the blacklist, we just suppress the "blacklist
entry changed!" warnings unless the offsetType is UTF-8.)

Change-Id: Idc3665c9c010105ce8cc5189462276fc7a2c655f