Link prefix fixes for Parsoid/PHP
Parsoid/PHP was missing the `D` regexp modifier when creating a link prefix
regexp from a character set, which caused the prefix to jump over newlines.

While auditing, noticed that the PHP prefix/trail regexps also contain the
`s` modifier (DOTALL).  However, node v6 doesn't support the `s` modifier,
and it appears that this is just used to handle the "bit after the link
trail" in the regexp -- the link trail for enwiki for example is:
```
   /^([a-z]+)(.*)$/sD
```

We actually strip the `(.*)$` part out in
WikiConfig.js::buildLinkNeighborRegex() and SiteConfig.php::linkTrailRegex()
so the `s` modifier should be unnecessary in both Parsoid/JS and Parsoid/PHP.

Bug: T236183
Change-Id: Ib4df6ed1dbe74aa9c9402bb9095e324b8becae42