Mostly use protocol-relative URLs for media
In Parsoid/JS and the standalone mode of Parsoid/PHP we always use
protocol-relative URLs for media, since the MediaWiki API (used by
Parsoid/JS and standalone Parsoid/PHP) always calls wfExpandUrl() on
the values it returns and we lose the original form.

For the integrated mode of Parsoid/PHP, we match the URL form used by
the legacy parser (since we can).  In Wikimedia production, this is
also protocol-relative, but it may diverge on third-party wikis with
different configurations.

Parser tests are a special case: the legacy parser runs parser tests
with a `http://` protocol.  Parsoid/JS uses protocol-relative (as
described above).  Likely when we get the integrated mode working
for Parsoid/PHP parser tests we'll have to tweak the configuration
to use protocol-relative mode there as well, since otherwise we'll
match the legacy parser and use `http://` by default.

Bug: T235217
Change-Id: If52d21b50cdbb466395ca64ac9877d992e19ce40