Return a 421 Misdirected Request if offsetType in incorrect
Parsoid/PHP puts an offsetType in its data-parsoid body.  If that
property is missing or not as expected, refuse to serialize, since this
is likely a misdirected request.

After Parsoid/JS is decommissioned, this check can be restored as a 406,
and possibly only applied if it was set, as before.

Bug: T237556
Change-Id: Ib4cfd535eca2688a2698b4492acf6e7ad5387ac9