Stop using DOMDataUtils in LC pass
Also, move the LC pass to after cleanup to reflect its true pure
transform nature.

When 554dfcf landed, checking data-parsoid (the call to
WTUtils.usesURLLinkSyntax(node)) merely left behind a .dataobject on a
node that would be dropped when serialized. After 7df5d97, it leaks a
data-object-id attribute. In other words, using DOMDataUtils in what are
supposed to be pure transformation on the output is fraught.

Bug: T234266
Change-Id: I72ae8f9b4c58adb54b7685deae3f04aaba5a4a61