Assert that template argument whitespace matching succeeds
The regexp is changed to avoid the excessive backtracking in some cases.

If the match runs up against the `pcre.backtrack_limit`, it fails, which
was being ignored.  That resulted in the matches array being empty and
php notices when trying to access it.  Here we prefer an early hard
failure.

Phan has a `UseReturnValuePlugin` that probably should have caught this
but it looks like it special cases `preg_match` and others.

Bug: T237103
Change-Id: I2d352cb0e35bd097e5ca282ba995984a1b107161