Use frame source instead of stringifying tokens
The page in the task was failing to stringify its pagebundle because the
offset mismatches from using stringified tokens when we reparsed the
extlink led to malformed UTF-8 characters.

This was noted as being suspect when introduced in 651f12c but at the
time, we probably didn't have the robust framing that makes this work.
JS was more robust to these errors because of UTF-16 offsets instead of
bytes.

This also throws exceptions if json_encode fails so that errors don't
go on suppressed.

Bug: T236867
Change-Id: Ic2e643389f483a6fcfe4c45dfec37b88911c9630