Stop unnecessarily wrapping templated targets when inTemplate
This normally wouldn't be an issue but when it occurs in a table tag
attribute position, it can mess with the logic in the attribute
expander.

Adding a test case for this is probably hard because of the current
limitations in parserTests doing its own template expansion, but it
probably involves a template in wikitext table tag syntax attribute
position that returns an unexpanded template with templated target.

The wikitext returned from expanding the template at the top this page
is as follows,

|- class="mergedtoprow"     [[:سانچو:Geobox2 line plain]] [[:سانچو:Geobox2 list]] [[:سانچو:Geobox2 list]]   {{ {{# if: |

Geobox2 list | Geobox 0}}

https://sd.wikipedia.org/w/rest.php/sd.wikipedia.org/v3/page/pagebundle/%DA%AF%D9%86%DA%AF%D8%A7/160038

Bug: T238457
Change-Id: Ic0079febcaa00008f77746cd8a3ec914d73ebf3f