html2wt selser: Return http 409 if the previous revision is not found
Note that we trigger this as an indication of failure to apply selective
serialization on the page.

Alternatively, we could fall back to regular serialization just like we
would if selser were not requested. In that scenario, we would return
serialized wikitext and the client will now encounter a failure when
trying to save the serialized wikitext.

But, http 409 is the correct response since we are not presuming that
the client cannot handle this scenario. We are narrowly confining our
response / error handling to what the client requested, and the client
could retry a non-selser request that will succeed and then provide
alternate handling in the lockdown scenario (ex: offer to the user to
create a different page which avoids losing the created content).

Bug: T238934
Change-Id: Iac17641159ab58614dbe34fd0729ac7d34dc3011