Assert that escaped text matching succeeds
The regexp there doesn't look like it ever made any sense since all it
wants to do is separate off the trailing newlines.

From inspection, we can get it to fail a match with "\r\r\n" but since
this is during serialization, our logging doesn't provide the actual
strings it saw.

Bug: T240091
Change-Id: I8383e52bf1d7526bb1bb41be637e5eb6e594490d