Ensure Sanitizer::sanitizeToken uses correct frame source text
This seems to have been an oversight when Token::getWTSource was changed
to take a frame -- we were passing the top-level frame unconditionally
instead of passing through the appropriate frame from the manager.  This
caused an embedded <ref> inside a <gallery> to use the wrong source text
which was then (one of the causes of) a UTF-8 corruption error.

Bug: T237318
Change-Id: Ia0128ab4a7462a257b72853e2d6473d1c8f5ce13