Test case for implicit newline insertion when expanding templates
Parsoid currently does this correctly in production because it calls
`expandTemplates` via the API and the necessary newlines are inserted
magically.  But the native expansion (used by parserTests) is incorrect,
because we expand the template arguments before knowing the correct
sol context, and so a template where the argument appears first (like
Template:Echo) has argument #1 expanded in non-SOL context.  When we
later splice in the argument tokens, they are still in non-SOL context
(ie, a literal `#` not a <li> token) and so adding an extra `\n` to
the token stream makes no difference.

If we eventually use the legacy preprocessor this will probably work
magically, but then we're committed to using a two-stage preprocess/parse
pipeline.  If we end up using a unified token stream, then we need to
somehow handle this case natively; probably be reparsing the argument
in SOL context.

Bug: T241150
Change-Id: I47a3cab8f6397abce5543cf121598108681ecdae