Refactor Parsoid::html2html into Parsoid::pb2pb
Use the PageBundle class to indicate when we're dealing with separated
HTML, data-parsoid, and data-mw, with a new PageBundle::toHtml()
method to convert it back into "unified" (one HTML DOM, inline
attributes) form.

This is a useful marker in the code --- otherwise it's hard to tell if
a variable named $html is "unified" HTML or just one part of a page
bundle).  (Internally, there's another format used, where the data
attributes have been loaded into a separate bag structure attached
to the HTML, but that is not exposed in our external interfaces.)

Similarly, create explicit Parsoid::pb2pb and Parsoid::html2html
endpoints so the input types are clear.  While we're at it, add a pair
of missing ContentUtils::convertOffsets()/ConvertOffsets::run() calls
so (a) the Env "request offset type" parameter is honored on input, if
provided, (b) the "current offset type" parameter in the Env is
initialized correctly, and (c) the output offset type is correctly
recorded in the page bundle by ConvertOffsets::run().

We don't expect to use "request offset type" in production, so (b) and
(c) are actually the important things here, just to ensure that the
Env is initialized correctly and the page bundle records the fact
that these are (almost always) "byte" offsets.

Change-Id: I4a5918d7cdf6e1abb3c806b07464bda8f6ae82b3