Use DOMDataUtils::getNodeData in MachineLanguageGuesser
PHP lacks a general-purpose hash table which can use arbitrary objects
(DOM nodes in this case) as keys.  We've tried using expando properties
directly on the PHP objects before, and found that the properties
spontaneously disappear when the original PHP object goes out of scope
and later a new PHP wrapper is created from scratch to wrap the same
underlying C object.  I had been using Node::getNodePath() to give a
string key which can be used in the table, but T241146 shows that this
can break: once LanguageConverter has started to insert new <span> tags
in Text nodes for round-trip information, the NodePath() of subsequent
<span> tags changes, which then causes crashes when the keys corresponding
to these aren't found.

Use the existing DOMDataUtils::getNodeData() framework to associate
information with Elements instead.  Since this only works on elements,
not Nodes, turn the traversal inside-out a little bit so we manually
compute and sum the results for Text node children of an Element.
In the process, simplify the complicated "shared array" logic of the
original Parsoid/JS version, since PHP gives us shared copy-on-write
arrays "for free".

Unfortunately, now that we're using the getNodeData() framework we
need to move language conversion earlier in the DOMPostProcessor
pipeline in order to ensure that the data-object-ids are properly
cleaned up and don't leak into the output.  Uses the refactored
Parsoid::pb2pb from I4a5918d7cdf6e1abb3c806b07464bda8f6ae82b3 to
facilitate loading/storing the data attributes.  Typically, the
`data-parsoid` attribute isn't supplied to the external html2html API,
but the `id` attributes which index the page bundles *are* included;
if we do this wrong we end up reassigning/scrambling the
`id`/`data-object-id` attributes.  Using the proper "page bundle"
interfaces even when no `data-parsoid` contents are provided helps
ensure that the `id`/`data-object-id` attributes are kept consistent,
and there's a test case in ParsoidTest demonstrating that we do
this correctly.

Bug: T241146
Change-Id: Ib460ccf1aba75b43db20d9df54a9aca4be9b26a8