Make ParsoidHandler::updateRedLinks() consistent with ::languageConversion
Use the pb2pb interface for both.

It's ok to load and store data attribs like this because, when they're
applied, the ids are left around and reused during storage.  That's
important because, unlike the variant case, redlink updated content will
be used for editing but RESTBase won't pass in or update the
data-parsoid based on these calls.

Followup-to: I1a583d096f204a5003daba3b43d34a86b5d2b4c1
Change-Id: I2a197bdb425815f6eb9aaba6c48ad84299d63ea4