Ensure Parsoid on scandium (RT test server) executes from git checkout
Parsoid-as-extension doesn't need to autoload its main composer.json
any more, since the main Parsoid codebase is already included in core
mediawiki as a library.

But when running on our round-trip-testing server, we want to explicitly
run Parsoid from a git checkout, *not* the library version of Parsoid.
Add a hack in tests/RTTestSettings.php to ensure that.

Bug: T240055
Change-Id: Ie336f69692317d7a714809b42e035fb188e66fc0