Rename extension to simply 'Parsoid'; update docs
Updated the description to reflect that fact that the extension *only*
implements the REST API, and (hopefully) will not be needed for LTS
(unless you want to expose the REST API).

The URL for the project has been changed to [[mw:Parsoid/PHP]], which
contains the install info for the PHP port of Parsoid; it is expected
that [[Parsoid/PHP]] will become a redirect to [[mw:Parsoid]] once we
release the next LTS and Parsoid/PHP becomes the 'official' way of
running Parsoid.

Also update instructions in extension/README.md to redirect to the wiki,
the canonical source of installation information.

Change-Id: I8a9ee8dc7631192da8e5bb56e6b828adf1982350