Remove exposure of KV objects by expanding ParsoidExtensionAPI
The KV object exposes Tokens and a bunch of other internals
which extensions have no need to worry about. If it were just
a plain old (k, v) object with some additional information,
we may have not have to make it opaque. In the future, we may
be able to convert the internal KV objects to a sanitized external
KV object.

But for now, the new API methods may actually lead to some code
simplification in extension implementations and so the opaqueness
may not be a bad thing after all.

This can be revisited based on feedback before we finalize this
API or at a later point in time.

Bug: T242746
Change-Id: I7ef696693159d965bf770e448671e4adbf065737