Update mocha api.js tests for Parsoid/PHP output
This should make it easier to catch bugs when refactoring.

Based on the output from 3ce3dee74 before refactoring began.

This would have caught the bug from T248646 fixed in 5b6c1aa since the
test "should get from a title and revision (wikitext)" fails on a3ee04c
returning a 302 instead of 200.

The tests can be run as described in T221174#5567849

Bug: T248646
Change-Id: Ia681e6469eff35c4bfa0cfb3b320f9817d31af78