Add fix for interwiki hrefs missing a $1 placeholder to extension/
An interwiki prefix with a url missing a $1 is added to
TestRunner::PARSER_TESTS_IWPS so that the fix in
ConfigUtils::computeInterwikiMap gets exercised.  Now, without the fix,
we'll actually get test failures.

However, the bug here is in the extension config which doesn't get
exercised in any of our test rigs.

Bug: T246701
Bug: T231568
Change-Id: I7d87f5246c44be698ff75585166915a43d4789f5