Run phan on code in extension/
We leverage the built-in mechanisms for using $MW_INSTALL_PATH to find
an installed mediawiki-core and adding that to the list of analyzed
phan directories.

The composer `phan` target does the "quick" phan of Parsoid code as an
independent codebase.

The new `phan-integrated` target does the "slow" phan, including the
extensions directory.

The `composer test` target still does only the 'quick' phan; we'll
add the integrated-mode phan to Jenkins as part of its usual set of
"tests done on extensions" (ie, run phan using `.phan/config.php`).

Change-Id: I4576a4a85cf1ac601dc5cea96bb42f593b8da183