Native ImageMap implementation
This is just a first pass that leaves quite a few FIXMEs.

The "imagemap_desc_types" can be localized, which isn't supported here
yet.

The legacy extension does ad hoc link parsing, which results in link
content not interpreting wikitext syntax.  This results in a known
difference to how it's done here, which just takes the text content of
the resulting dom.

For now, this implementation punts on adding the magnifying / info
links to avoid the div soup.  For media thumbs in general, Parsoid uses
a css class to add them.  Perhaps a similar approach can be taken here.

The map gets added after the image but inside the figure, which is
different to how the legacy parser does it.  Do browsers mind?

The implementation leaves the images as inline, if they were written
that way, whereas the legacy parser may force block unnecessarily.

Some new tests are added to capture various edge and error cases, since
the test suite was woefully lacking.

Bug: T94793
Depends-On: I6ff81a01207e2734090c626b177e5f4d10bb6d61
Change-Id: Ia91a10d178fc3b3b9d22a1d49210bde82e981ee1