npm update mkdirp --depth 5
Resolves 5 vulns, mostly in dev deps

We can dig into the one in bunyan, which is a dependency of
service-runner.  It's in a sub-dep, mv, which is an optional dependency
so it was never installed in the deploy repo.

mv is only used with the RotatingFileStream and vuln in minimist would
only be relevant if the attacker had control over the arguments to
mkdirp but that's not likely to be the case for the log file name.

https://www.npmjs.com/advisories/1179

Bug: T248208
Change-Id: I1b13707241f5cff57e0e446d65b13590254b08d6