Upgrade to request@2.87.0
Resolves 3 vulns

Two are in a dependency, hawk, which gets upstreamed into request in the
new version so there's no path to upgrading that independently.

hawk is only used when making requests to servers using
http authentication but parsoid/js hasn't been making use of that
feature.

https://github.com/request/request#http-authentication

https://www.npmjs.com/advisories/720
https://www.npmjs.com/advisories/1464

The other is in another dependency, extend.  The attack looks like it
requires control over the source objects to extend with, but the uses in
request seem mainly to be about extending options whose keys would have
been set in parsoid/js source.

https://www.npmjs.com/advisories/996

Bug: T248208
Change-Id: I4313bfbe3d084c058d707f0b375076be378d5c3e