Two minor tweaks to our autogenerated documentation
The heading in CODE_OF_CONDUCT.md seems to be necessary to make it
actually show up on docs.wikipedia.org.

And for some reason "Diff tools" is showing up as the default description
for the Wikimedia\Parsoid namespace.  Hopefully putting the
`src/Utils/Diff.php` file (the last unported JavaScript file, AFAICT)
in the proper Utils namespace will fix the description of Wikimedia\Parsoid.

Change-Id: Ib67bbac4aea0f78ae928f3331796f16af3bc60e5