Move armoring french spaces to a post-processing pass
The isDisplayHack flag in data-parsoid is no longer used or necessary,
so gets removed.

A number of tests started failing wt2wt because a mw:DisplaySpace was no
longer falsely being emitted, which had the side effect of preserving
whitespace while roundtripping.  These tests were converted to
`parsoid=wt2html,html2html` since they aren't expected to pass as
written.

Despite what it says, the test
"T71950: 2. Put nowiki as close to cause as possible, after ' :'"
was failing because, in f997f0b, all it did was skip the start and end
tokens and leave behind the &nbsp;.  In order to work, it needed to skip
all three and buffer a plain space characer as in hasWikitextTokens.
Thankfully, checks for isDisplayHack can be removed from the wikitext
escaping since they're no longer produced in the tokenizer.

T254501 was filed to remove the "mw:DisplaySpace"s when in a
preprocessing pass when serializing so that dealing with them isn't
strewn out all over the codebase.

T254502 was filed to see if we can remove the mw:Placeholder from these
spans.

T254500 was filed to update the spec to include mw:DisplaySpace and its
expectations.

Bug: T197879
Change-Id: Ia164dd1318d45924aa965919e7939c6f817f5d0d