Bump version of Remex to v2.2.0 (bump zest and alea as well)
This version of Remex has PHP 7.4 compatibility fixes and drops support
for PHP 7.0, PHP 7.1, and HHVM.  THere are also some minor bug fixes,
but not in code used by Parsoid.

While we're at it, bump the versions of zest and alea so none of them
have dev-dependencies on Remex 2.1.0 either.  Not strictly necessary,
but we might as well.

Bug: T250935
Depends-On: I6bcb443b00295911a12b97830bd67c52c10dc469
Change-Id: Ifa368f6bc3dc8ab5c4500d5964b5fedaceffe92f