Add caption to always suppressing
This matches the dependent core commit and accounts for the fact that
Parsoid relies on the treebuilder to close table cells.

The test added in 18a98af is no longer relevant now that the paragraph
wrapping that led to the case has been suppressed.  The entire patch
can't be reverted since it seems to be covering a bunch of other
untested cases as well.

This was discussed in the review,
https://gerrit.wikimedia.org/r/c/mediawiki/services/parsoid/+/476053/3/tests/parserTests.txt#31161

Bug: T210647
Depends-On: I8ccefd69d47dca740f50924b235dffa3873d1f99
Change-Id: Idf3081ca118f41bf2506dfe23aad020b09ed732c