Apply the inline_breaks to all the expressions in path
Throughout the grammar, inline_breaks are used to prevent matching on
all the expressions that succeed them in a rule definition.  If not, it
sort of defeats the purpose of breaking for a higher precedent rule
that's been pushed.

Since 6df94a3, "url" and "autourl" have sort of been the ugly ducklings
in that they only applied it to character class in the first expression.

For the sake of consistency, and to bring things in line with the
assumed intended use, we apply the inline_breaks to all the subsequent
expressions.

This is a change in behaviour but it's likely to be for the good.

Change-Id: I767cee71002f30539fa8b006c1a19f9b873648be