Stop unwrapping mw:DisplaySpace spans in parserTest results
This was added in 4d6b87b, for T94509.  At the time it was necessary
because of all the falsely emitted spans that were breaking the
definition list tests.  That was resolved in the prior patch to this so
it's no longer necessary.

Instead, update a few parserTests with html/parsoid sections where we
actually expect the mw:DisplaySpace to be found.

Change-Id: I66752cb408afb048d1c087ee502ffd58b94d9211