Revert most of c7403da
Now that I438cd9a113c15b1eb7bf15c787b4da15728bc12d is merged and
released.

The unrelated type annotations that were added can stay.

The nice thing about reverted implementation is that the transclusion
shadow meta is only inserted at the end of a run, rather than after the
first token.  See the follow up for trying to reproduce that in this
model.

Change-Id: I0bc99df59987779c502dc361bf97218f019fc703