Remove unused Parsoid::html2html
When this method was added, the current pb2pb transformations were seen
as pure HTML post-processing steps, and didn't require the
data-attributes be loaded.

In 1641a20, the data-attributes were made use of in the variant
hydration and so a Parsoid::pb2pb was added in 512608d.  For
consistency, the redlinks pass was converted to using the same endpoint
in I2a197bdb425815f6eb9aaba6c48ad84299d63ea4 making html2html redundant.

Change-Id: Ida903fcc65df0a55ccc9e466b1b141445555f83e