DRY out default ParsoidSettings into a single location
We need to fetch $wgParsoidSettings from the config object, catch if it
throws, and set up defaults.  This was done inconsistently in multiple
places previously, causing exceptions if your LocalSettings.php left
$wgParsoidSettings unset.  Use a ParsoidSetttings service so that this
code is done in a single place with a single set of defaults.

Currently we default to 'useSelser' => true, but this could be changed
for LTS if performance is more important than dirty diffs.  We also updated
the extension description to clarify that (for the time being) the Parsoid
extension is required even if you're not using RESTbase.

Bug: T248343
Change-Id: I435bdbba7d400b4226c3296cbda238dc0954edbc